+++ doc/changelog.md
@@ -1,5 +1,6 @@
 2018/11/24
 * added option to mount multiple folders
+* added backup restore for mysql databases on start/stop of VM
 
 2018/11/21
 * initial commit with basic vagrant file